Modularize ProfileStackNavigator
